+++ core/admin.py
@@ -1,3 +1,6 @@
 from django.contrib import admin
 
+from .models import Chamado
+
 # Register your models here.
+admin.site.register(Chamado)
