+++ core/static/main.js
